Fix window reload
