+++ src/modules/display-tasks.js
@@ -1,5 +1,7 @@
-/* Display tasks */
-
+// eslint-disable-next-line import/no-cycle
+import {
+  addTask, checkBoxesStatus, deleteTask, edit,
+} from './functions.js';
 import { store, taskArray } from './store.js';
 
 const displayTasks = () => {
@@ -28,6 +30,57 @@ const displayTasks = () => {
     `;
       store();
     });
+
+  /* Add */
+  const addBtn = document.querySelector('.fa-right-from-bracket');
+  addBtn.addEventListener('click', (e) => {
+    addTask(e);
+  });
+
+  /* Delete */
+  const trashCans = document.querySelectorAll('.delete-icon');
+  for (let i = 0; i < trashCans.length; i += 1) {
+    trashCans[i].addEventListener('click', (e) => {
+      displayTasks();
+      deleteTask(e);
+    });
+  }
+
+  /* Delete completed */
+  const deleteCompleted = document.querySelector('.completed-text');
+  const allCheckBoxes = document.querySelectorAll('input[type=checkbox]');
+
+  function clearCompleted(elem) {
+    return elem.completed !== true;
+  }
+
+  deleteCompleted.addEventListener('click', () => {
+    const data = taskArray.filter(clearCompleted);
+    localStorage.setItem('taskInput', JSON.stringify(data));
+    window.location.reload();
+  });
+
+  for (let i = 0; i < allCheckBoxes.length; i += 1) {
+    allCheckBoxes[i].addEventListener('click', (e) => {
+      checkBoxesStatus(e);
+    });
+  }
+
+  /* Edit */
+  const taskTexts = document.querySelectorAll('p');
+  for (let i = 0; i < taskTexts.length; i += 1) {
+    taskTexts[i].addEventListener('click', (e) => {
+      edit(e);
+    });
+  }
+
+  /* Delete all */
+  const deleteAll = document.querySelector('.reload-icon');
+  deleteAll.addEventListener('click', () => {
+    taskArray.splice(0, taskArray.length);
+    store();
+    displayTasks();
+  });
 };
 // fa-regular  fa-trash-can delete-icon
 export default displayTasks;
